requirements.txt
streamlit
openai
simpleaichat
ulid-py
deta